fix(levoit): resolve compiler warnings in component build
Clears every warning the Levoit component emits during an ESP-IDF build
except the -Wswitch one in on_number_command, which cannot be silenced
without deciding whether white_noise_min should reach the MCU. No
behavior change.

- types.h: make command_type_to_string inline instead of static. As a
  static function defined in a header it was reported unused by each of
  the nine translation units that include types.h without calling it,
  which accounted for most of the warning output.
- levoit.cpp: cast uint32_t log arguments to unsigned in four ESP_LOGx
  calls. uint32_t is unsigned long on xtensa, so %u mismatched its
  argument type. Matches the existing (unsigned) cast style used
  elsewhere in the component.
- decoder.cpp: parenthesize the && operands inside the || in the
  Core300S/400S status ptype test. Grouping is unchanged; this only
  states it explicitly.

+++ components/levoit/levoit.cpp
@@ -174,7 +174,7 @@ namespace esphome
                 const char *opt = fan_operating_mode_to_option_(value);
                 if (opt == nullptr)
                 {
-                    ESP_LOGW(TAG, "publish_select: invalid fan mode %u", value);
+                    ESP_LOGW(TAG, "publish_select: invalid fan mode %u", (unsigned)value);
                     return;
                 }
                 if (std::find(options.begin(), options.end(), opt) == options.end())
@@ -190,7 +190,7 @@ namespace esphome
             if (value >= options.size())
             {
                 ESP_LOGW(TAG, "publish_select: invalid index %u for type %d (options=%u)",
-                         value, (int)type, (unsigned)options.size());
+                         (unsigned)value, (int)type, (unsigned)options.size());
                 return;
             }
             const std::string &opt = options[value];
@@ -664,13 +664,13 @@ namespace esphome
             
             // Restore saved values or initialize to 0
             if (pref_used_cadr_.load(&used_cadr_)) {
-                ESP_LOGI(TAG, "Restored used_cadr: %u m³", used_cadr_);
+                ESP_LOGI(TAG, "Restored used_cadr: %u m³", (unsigned)used_cadr_);
             } else {
                 used_cadr_ = 0;
                 ESP_LOGI(TAG, "Initialized used_cadr to 0");
             }
             if (pref_total_runtime_.load(&total_runtime_)) {
-                ESP_LOGI(TAG, "Restored total_runtime: %u hours", total_runtime_);
+                ESP_LOGI(TAG, "Restored total_runtime: %u hours", (unsigned)total_runtime_);
             } else {
                 total_runtime_ = 0;
                 ESP_LOGI(TAG, "Initialized total_runtime to 0");
